+++ R/amt.R
@@ -1,21 +1,17 @@
 #' @importFrom dplyr arrange filter group_by select summarize summarise ungroup rename
-#' @importFrom dplyr distinct ungroup mutate mutate_at pull %>% bind_rows select_vars n
+#' @importFrom dplyr distinct ungroup mutate mutate_at pull bind_rows select_vars n
 #' @importFrom graphics plot legend lines par points
 #' @importFrom grDevices adjustcolor
-#' @importFrom lubridate hours minutes seconds now days weeks
-#' @importFrom magrittr %>%
+#' @importFrom lubridate hours minutes seconds now days weeks %within%
 #' @importFrom stats coef dgamma median na.omit qgamma quantile runif sd var formula reformulate qnorm
 #' @importFrom tidyr nest unnest
 #' @importFrom tibble tibble tribble is_tibble as_tibble
 #' @importFrom utils data head tail
 #' @importFrom sf NA_crs_
-#' @importFrom sp CRS
 #' @importFrom stats dexp qexp terms predict
 #' @importFrom methods as is
 #' @importFrom purrr map map_int
-#' @importFrom Rcpp evalCpp
 #' @importFrom rlang quo quos enquo quo_name
-#' @importFrom raster raster
 #' @import survival
 #' @import Rdpack
 #' @export
@@ -26,12 +22,12 @@ methods::setOldClass(c("random_points", "tbl_df", "tbl", "data.frame"))
 methods::setOldClass(c("steps", "tbl_df", "tbl", "data.frame"))
 methods::setOldClass(c("random_steps", "tbl_df"))
 
-utils::globalVariables(c("burst_", "step_id_", "t_", "case_", ".data",
-                         "ts", "sl_", "xs", "ys", "x_", "y_")) # to omit CRAN notes
-
-#' @useDynLib amt
-
-"_PACKAGE"
+utils::globalVariables(
+  c("burst_", "step_id_", "t_", "case_", ".data",
+    "ts", "sl_", "xs", "ys", "x_", "y_", "x1_", "y1_",
+    "y0_", "x0_", "y2_", "x2_", "abs.dir2", "rel.dir", "ta_",
+    "step_id_1", "abs.dir1", "x", "y", ":=", ".I", "t2_",
+    "linestring_id")) # to omit CRAN notes
 
 
 # exports from other packages ---------------------------------------------
@@ -61,8 +57,6 @@ dplyr::mutate
 #' @export
 dplyr::pull
 #' @export
-sp::CRS
-#' @export
 lubridate::hours
 #' @export
 lubridate::minutes
@@ -89,8 +83,4 @@ tidyr::unnest
 #' @export
 tibble::tibble
 #' @export
-magrittr::`%>%`
-#' @export
 survival::Surv
-#' @export
-raster::raster