+++ R/extract_covariates.R
@@ -3,31 +3,24 @@
 #' Extract the covariate values at relocations, or at the beginning or end of
 #' steps.
 #' @template track_xy_star_steps
-#' @param covariates `[RasterLayer,RasterStack,RasterBrick]` \cr The
+#' @param covariates `[SpatRaster]` \cr The
 #'   (environmental) covariates. For `extract_covariates_var_time` the argument
 #'   `covariates` need to have a `z`-column (i.e. the time stamp).
 #' @param where `[character(1)="end"]{"start", "end", "both"}` \cr For `steps`
 #'   this determines if the covariate values should be extracted at the
 #'   beginning or the end of a step. or `end`.
-#' @param ... Additional arguments passed to `raster::extract()`.
+#' @param ... Additional arguments passed to `terra::extract()`.
 #' @return A `tibble` with additional columns for covariate values.
 #' @name extract_covariates
 #' @export
 #' @examples
 #' data(deer)
-#' data(sh_forest)
+#' sh_forest <- get_sh_forest()
 #' mini_deer <- deer[1:20, ]
-#' mini_deer %>% extract_covariates(sh_forest)
-#' mini_deer %>% steps %>% extract_covariates(sh_forest)
-#' mini_deer %>% steps %>% extract_covariates(sh_forest, where = "start")
+#' mini_deer |> extract_covariates(sh_forest)
+#' mini_deer |> steps() |> extract_covariates(sh_forest)
+#' mini_deer |> steps() |> extract_covariates(sh_forest, where = "start")
 #'
-#' # Buffer
-#' mini_deer %>% extract_covariates(sh_forest) # no buffer
-#' # The command buffer can be used, to buffer each point together with a
-#' # function to summarize the results.
-#' mini_deer %>% extract_covariates(sh_forest, buffer = 10, fun = mean)
-#' # This can also be a use-specified function.
-#' mini_deer %>% extract_covariates(sh_forest, buffer = 100, fun = function(x) length(x))
 
 extract_covariates <- function(x, ...) {
   UseMethod("extract_covariates", x)
@@ -48,36 +41,30 @@ extract_covariates.random_points <- function(x, covariates, ...) {
 #' @export
 #' @rdname extract_covariates
 extract_covariates.steps_xy <- function(x, covariates, where = "end", ...) {
-  if (class(covariates) %in% paste0("Raster", c("Layer", "Stack", "Brick"))) {
-    if (where == "both") {
-      x_start <- raster::extract(covariates, as.matrix(x[, c("x1_", "y1_")]),
-                                 df = TRUE, ...)[, -1, drop = FALSE]
-      names(x_start) <- paste0(names(x_start), "_start")
-      x_end <- raster::extract(covariates, as.matrix(x[, c("x2_", "y2_")]),
-                               df = TRUE, ...)[, -1, drop = FALSE]
-      names(x_end) <- paste0(names(x_end), "_end")
-      x_all <- cbind(x_start, x_end)
-      x[names(x_all)] <- as.data.frame(x_all)
-    } else {
-      x[names(covariates)] <- if (where == "end") {
-        as.data.frame(raster::extract(covariates, as.matrix(x[, c("x2_", "y2_")]), ...))
-      } else if (where == "start") {
-        as.data.frame(raster::extract(covariates, as.matrix(x[, c("x1_", "y1_")]), ...))
-      }
-    }
-    x
+
+  validate_covars(covariates)
+  if (where == "both") {
+    x_start <- terra::extract(covariates, as.matrix(x[, c("x1_", "y1_")]), ...)
+    names(x_start) <- paste0(names(x_start), "_start")
+    x_end <- terra::extract(covariates, as.matrix(x[, c("x2_", "y2_")]), ...)
+    names(x_end) <- paste0(names(x_end), "_end")
+    x_all <- cbind(x_start, x_end)
+    x[names(x_all)] <- x_all
   } else {
-    stop("no raster")
+    x[names(covariates)] <- if (where == "end") {
+      terra::extract(covariates, as.matrix(x[, c("x2_", "y2_")]), ...)
+    } else if (where == "start") {
+      terra::extract(covariates, as.matrix(x[, c("x1_", "y1_")]), ...)
+    }
   }
+  x
 }
 
 extract_covar_base <- function(x, covars, ...) {
-  if (class(covars) %in% paste0("Raster", c("Layer", "Stack", "Brick"))) {
-    x[names(covars)] <- as.data.frame(raster::extract(covars, x[, c("x_", "y_")], ...))
-    x
-  } else {
-    stop("no raster")
-  }
+  covars <- validate_covars(covars)
+  x[names(covars)] <- terra::extract(covars, x[, c("x_", "y_")],
+                                     ID = FALSE, ...)
+  x
 }
 
 
@@ -90,11 +77,11 @@ extract_covar_base <- function(x, covars, ...) {
 #' @examples
 #' \donttest{
 #' # Illustration of extracting covariates along the a step
-#' mini_deer %>% steps() %>% random_steps() %>%
-#'   extract_covariates(sh_forest) %>% # extract at the endpoint
-#'   mutate(for_path = extract_covariates_along(., sh_forest))  %>%
+#' mini_deer |> steps() |> random_steps() |>
+#'   extract_covariates(sh_forest) |> # extract at the endpoint
+#'   (\(.) mutate(., for_path = extract_covariates_along(., sh_forest)))()  |>
 #'   # 1 = forest, lets calc the fraction of forest along the path
-#'   mutate(for_per = purrr::map_dbl(for_path, ~ mean(. == 1)))
+#'   mutate(for_per = purrr::map_dbl(for_path, function(x) mean(x == 1)))
 #' }
 
 extract_covariates_along <- function(x, ...) {
@@ -104,17 +91,12 @@ extract_covariates_along <- function(x, ...) {
 #' @export
 #' @rdname extract_covariates
 extract_covariates_along.steps_xy <- function(x, covariates, ...) {
-  if (class(covariates) %in% paste0("Raster", c("Layer", "Stack", "Brick"))) {
-    wkt <- with(x, paste0("LINESTRING (", x1_, " ", y1_, ",", x2_, " ", y2_, ")"))
-    ll <- sf::st_as_sfc(wkt)
-    # Remove dependency on velox
-    # v <- velox::velox(covariates)
-    # l2 <- v$extract(sp = ll)
-    l2 <- raster::extract(covariates, sf::as_Spatial(ll), ...)
-    return(l2)
-  } else {
-    stop("covariates: need to be a Raster*.")
-  }
+  covariates <- validate_covars(covariates)
+
+  wkt <- with(x, paste0("LINESTRING (", x1_, " ", y1_, ",", x2_, " ", y2_, ")"))
+  ll <- sf::st_as_sfc(wkt)
+  l2 <- terra::extract(covariates, terra::vect(ll), ...)
+  stats::setNames(split(l2[, -1], l2[, 1]), NULL)
 }
 
 # Extract covariates varying time -----------------------------------------
@@ -149,6 +131,8 @@ extract_covariates_var_time.steps_xyt <- function(
   x, covariates, when = "any", max_time, name_covar = "time_var_covar",
   where = "end", ...) {
 
+  covariates <- validate_covars(covariates)
+
   if (where == "start") {
     x[name_covar] <- extract_covar_var_time_base(
       cbind(x$x1_, x$y1_),
@@ -173,15 +157,15 @@ extract_covar_var_time_base <- function(
   xy, t, covariates, when = "any",
   max_diff, ...) {
 
-  if (is.null(raster::getZ(covariates))) {
-    stop("Covariates do not have a Z column.")
+  if (is.null(terra::time(covariates))) {
+    stop("Covariates do not have a time stamp. Use `terra::time()` to assign one.")
   }
 
   if (!is(max_diff, "Period")) {
     stop("`max_diff` is not of class `Period`.")
   }
   max_diff <- lubridate::period_to_seconds(max_diff)
-  t_covar <- as.numeric(as.POSIXct(raster::getZ(covariates)))
+  t_covar <- as.numeric(as.POSIXct(terra::time(covariates)))
   t_obs <- as.numeric(as.POSIXct(t))
 
   # Fun to find closest point
@@ -203,10 +187,17 @@ extract_covar_var_time_base <- function(
   }
 
   wr <- sapply(t_obs, function(x) which_rast(x - t_covar, when, max_diff))
-  ev <- raster::extract(covariates, cbind(xy), ...)
+  ev <- terra::extract(covariates, cbind(xy), ...)
   cov_val <- ev[cbind(seq_along(wr), wr)]
   return(cov_val)
 }
 
 
+raster_to_terra <- function(x) {
+  if (class(x) %in% paste0("Raster", c("Layer", "Stack", "Brick"))) {
+    terra::rast(lapply(list(x), function(r) terra::rast(r)))
+  } else {
+    "No raster or SpatRast provided"
+  }
+}