+++ R/fit_clogit.R
@@ -17,7 +17,14 @@
 #' @export
 
 fit_clogit <- function(data, formula, more = NULL, summary_only = FALSE, ...) {
+
+  # Ensure strata is provided
+  if (!any(grepl(pattern = "^strata\\(.+\\)$", attr(terms(formula), "term.labels")))) {
+    stop("No strata is provided, please make sure the formula includes a strata.")
+  }
+
   m <- survival::clogit(formula, data = data, ...)
+
 
   if (summary_only) {
     m <- list(model = broom::tidy(m),
@@ -37,7 +44,10 @@ fit_clogit <- function(data, formula, more = NULL, summary_only = FALSE, ...) {
 
 #' @export
 coef.fit_clogit <- function(object, ...) {
-  stats::coef(object$model, ...)
+  if (is(object$model, "list"))
+    return(object$coefficients)
+  else
+    stats::coef(object$model, ...)
 }