+++ inst/tinytest/test_log_rss.R
@@ -1,14 +1,14 @@
 library(amt)
 data(deer)
-data("sh_forest")
-d <- deer[1:50, ] %>% random_points() %>% extract_covariates(sh_forest) %>%
+sh_forest <- get_sh_forest()
+d <- deer[1:50, ] |> random_points() |> extract_covariates(sh_forest) |>
   mutate(w = ifelse(case_, 1, 1000))
 
-f1 <- glm(case_ ~ sh.forest, data = d, family = binomial())
-f2 <- fit_logit(d, case_ ~ sh.forest)
+f1 <- glm(case_ ~ forest, data = d, family = binomial())
+f2 <- fit_logit(d, case_ ~ forest)
 
-x1 <- data.frame(sh.forest = 0)
-x2 <- data.frame(sh.forest = 1)
+x1 <- data.frame(forest = 0)
+x2 <- data.frame(forest = 1)
 
 rss1 <- log_rss(f1, x1, x2)
 rss2 <- log_rss(f2, x1, x2)
@@ -18,20 +18,21 @@ expect_equal(rss1, rss2)
 # Test `check_factors`
 
 data("amt_fisher")
+amt_fisher_covar <- get_amt_fisher_covars()
 
 # Prepare data for RSF
-rsf_data <- amt_fisher %>%
-  filter(name == "Lupe") %>%
-  dplyr::slice_sample(n = 250) %>%
-  make_track(x_, y_, t_) %>%
-  random_points() %>%
-  extract_covariates(amt_fisher_covar$elevation) %>%
-  extract_covariates(amt_fisher_covar$popden) %>%
-  extract_covariates(amt_fisher_covar$landuse) %>%
+rsf_data <- amt_fisher |>
+  filter(name == "Lupe") |>
+  dplyr::slice_sample(n = 250) |>
+  make_track(x_, y_, t_) |>
+  random_points() |>
+  extract_covariates(amt_fisher_covar$elevation) |>
+  extract_covariates(amt_fisher_covar$popden) |>
+  extract_covariates(amt_fisher_covar$landuse) |>
   mutate(lu = factor(landuse))
 
 # Fit RSF without factor
-m1 <- rsf_data %>%
+m1 <- rsf_data |>
   fit_rsf(case_ ~ elevation + popden)
 
 # data.frame of x1s
@@ -45,7 +46,7 @@ x2 <- data.frame(elevation = mean(rsf_data$elevation),
 expect_null(amt:::check_factors(m1$model, x1, x2))
 
 # Fit RSF with factor
-m2 <- rsf_data %>%
+m2 <- rsf_data |>
   fit_rsf(case_ ~ lu + elevation + popden)
 
 # data.frame of x1s